requirements.txt
boto3==1.5.8
botocore==1.8.22
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.1
django-allauth==0.36.0
django-bulma==0.5.1
django-categories==1.6
-e git://github.com/cpkthompson/django-category.git@4a2729ce9f6567a5cad1c3c0223575bf4af86ac8#egg=django_category
django-heroku==0.3.1
django-js-asset==1.1.0
django-mptt==0.9.0
django-object-tools==1.11.0
django-storages==1.6.5
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
oauthlib==2.1.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.11.0
unicode-slugify==0.1.3
Unidecode==1.0.22
urllib3==1.23
whitenoise==3.3.1
